Create base test fixture which seeds test data for UDF tests.
